20230901 리뷰 전체 조회 기능 추가

+++ src/main/java/com/example/jj_book/dto/ReviewItemDto.java
@@ -9,7 +9,7 @@ import lombok.Setter;
 public class ReviewItemDto {
 
     private String itemNm;
-    private int rating; //리뷰별점
+    private Long rating; //리뷰별점
     private String contents; //리뷰내용
     private String imgUrl; //리뷰 이미지 경로
